pyproject.toml
[project]
name = "coderetrx"
version = "0.1.16"
description = "Library for Code Analysis via Static Analysis and LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.3.0",
    "git-python>=1.0.3",
    "httpx>=0.28.1",
    "ignore-python>=0.2.0",
    "json-repair>=0.46.0",
    "nanoid>=2.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-git>=2018.2.1",
    "tomli-w>=1.2.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-language-pack>=0.7.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["coderetrx*"]
exclude = []
namespaces = false

[dependency-groups]
dev = [
    "httpx[socks]>=0.28.1",
    "ipython>=9.2.0",
    "pytest>=8.3.5",
    "rich>=14.0.0",
]

[project.optional-dependencies]
stats = ["tiktoken>=0.9.0"]
builtin-impl = [
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "openai>=1.0.0",
]
cli = ["typer>=0.16.0"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"coderetrx" = ["**/*.scm", "**/*.py"]
